+++ backend/src/main/java/com/picook/domain/admin/auth/repository/AdminUserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface AdminUserRepository extends JpaRepository<AdminUser, Long> {
+public interface AdminUserRepository extends JpaRepository<AdminUser, Integer> {
 
     Optional<AdminUser> findByEmail(String email);
